Acceleration Appears to Work
The acceleration section of the code appears to work correctly.
After the few cycles through the main loop, the program appears to enter
a loop in the PASM section of code which continuously steps the motor.
The motor appear to turn at full speed when caught in this loop. There
is an occasional "stutter" of the motor while in this loop. I think this
stutter occurs during rollover of the main counter but I'm not sure if
the stutter is related to the system counter or not.