EXECUTEX Sort of Works
The program EXECUTEX will open a file from the SD card and move the
stepper motors based on the file data. The motion from the steppers
appears to be correct but I can only navigate the menus with the serial
terminal. The joystick and button control doesn't appear to work
correctly. The main trouble appears to be with the button.