Spiral Kind of Works
The MathTestSpiral program works reasonably well. It's slower than I
think it should be but I'm not sure about this.
The object "TestMotorX" appears to work correctly and the object
"TestMotor" does not compile.
I'll save these version as is for now and then begin adding the circle
code to the MotorControlX object but name it "MotorControl".